Projects Page We will reorganize this page as follows
+ The top section will contain the same information as the supporter listing pages.
+ The counters and icons will be moved to appear next to the 3 column headers (Cash Donors, etc)
+ The names of contributing people will be aligned underneath the column titles

If the visitor is not logged in, then we can assume that they will have come from a README type page and so we will include the following instruction below the header bar …

+“Welcome to I Love Open Source. You may make a donation by clicking below. If you sign in , you can add an I Love Open Source page to your site as well, showing which Open Source projects you use and helping to get recognition for those projects.” [Signin with Github]

Once the user has signed in, show into.

Supporters Page

+ We will add a “Share” button next to the Preview button. When the user clicks on it, a section should expand out, pushing down the rest of the content. In this section, the following should appear

An instruction text “Please copy and paste the Markdown or HTML link into your site. This is very important as it publicly expresses your support and encourages others to do so as well.”
Embed Link  a) Markdown and b) HTML that shows the ILOS icon and links to the User Profile page